When coming offline to online automatically,  Dropbox files are updated.
The onboard library files are not disturbed. and so some synchronisation is left to the tnext cold start.
pathCount.a has been expanded to record the filenames of the loaded Dropbox files for comparison with the database files. Variable wasOffline is used to identify that the loading has occured after the internet connection has been restored.