Offline database operation
this.online is set to the function isOnline() in create, but on the first execution becomes a boolean.
Need this.online to be set on entry to determine if database is an option
Establish the initial state of this.online by the value of  navigator.onLine

The database is available if HTML5 database is available, so tested for the existence of the openDatabse  function.